requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
PyAudio==0.2.13
python-decouple==3.8
pyttsx3==2.90
requests==2.30.0
SpeechRecognition==3.10.0
urllib3==2.0.2
